# docs/reference/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "cookie": {
        "in": "cookie",
        "name": "nexora_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "Nexora Platform Core API",
    "version": "1"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/password": {
      "put": {
        "operationId": "AuthenticationController_changeAuthenticatedPassword",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "currentPassword": {
                    "maxLength": 128,
                    "minLength": 1,
                    "type": "string",
                    "writeOnly": true
                  },
                  "newPassword": {
                    "maxLength": 128,
                    "minLength": 15,
                    "type": "string",
                    "writeOnly": true
                  }
                },
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password changed, existing sessions revoked, and current session rotated."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Change the current user password",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/v1/auth/password-reset-requests": {
      "post": {
        "operationId": "AuthenticationController_requestPasswordResetLink",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 254,
                    "type": "string"
                  }
                },
                "required": [
                  "email"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "The request is accepted regardless of account existence."
          }
        },
        "summary": "Request a password reset link",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/v1/auth/password-resets": {
      "post": {
        "operationId": "AuthenticationController_confirmPasswordReset",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "newPassword": {
                    "maxLength": 128,
                    "minLength": 15,
                    "type": "string",
                    "writeOnly": true
                  },
                  "token": {
                    "maxLength": 43,
                    "minLength": 43,
                    "type": "string",
                    "writeOnly": true
                  }
                },
                "required": [
                  "token",
                  "newPassword"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password replaced and every existing session revoked."
          }
        },
        "summary": "Replace a password using a reset token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/v1/auth/registrations": {
      "post": {
        "operationId": "AuthenticationController_register",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "displayName": {
                    "maxLength": 100,
                    "minLength": 1,
                    "type": "string"
                  },
                  "email": {
                    "format": "email",
                    "maxLength": 254,
                    "type": "string"
                  },
                  "organizationName": {
                    "maxLength": 120,
                    "minLength": 1,
                    "type": "string"
                  },
                  "password": {
                    "maxLength": 128,
                    "minLength": 15,
                    "type": "string",
                    "writeOnly": true
                  },
                  "workspaceName": {
                    "maxLength": 120,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password",
                  "displayName",
                  "organizationName",
                  "workspaceName"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Account created and session cookie issued."
          }
        },
        "summary": "Create an account and its initial organization and workspace",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/v1/auth/session": {
      "delete": {
        "operationId": "AuthenticationController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Current session revoked."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Revoke the current session",
        "tags": [
          "Authentication"
        ]
      },
      "get": {
        "operationId": "AuthenticationController_currentSession",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current authenticated session."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Resolve the current user and trusted active workspace",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/v1/auth/session/workspace": {
      "put": {
        "operationId": "AuthenticationController_selectWorkspace",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "workspaceId": {
                    "format": "uuid",
                    "type": "string"
                  }
                },
                "required": [
                  "workspaceId"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Active workspace switched and the opaque session cookie rotated."
          },
          "403": {
            "description": "The requested workspace is not available to this user."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Switch the active workspace for this session",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/v1/auth/session/workspaces": {
      "get": {
        "operationId": "AuthenticationController_availableWorkspaces",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current workspace memberships."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List workspaces available to the current user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/v1/auth/sessions": {
      "delete": {
        "operationId": "AuthenticationController_logoutEverywhere",
        "parameters": [],
        "responses": {
          "204": {
            "description": "All user sessions revoked."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Revoke every session for the current user",
        "tags": [
          "Authentication"
        ]
      },
      "post": {
        "operationId": "AuthenticationController_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 254,
                    "type": "string"
                  },
                  "password": {
                    "maxLength": 128,
                    "minLength": 1,
                    "type": "string",
                    "writeOnly": true
                  },
                  "workspaceId": {
                    "format": "uuid",
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Session created and cookie issued."
          },
          "409": {
            "description": "Valid credentials require an explicit choice from the returned workspaces."
          }
        },
        "summary": "Authenticate and create a new session",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "servers": [],
...
}